Java changes to make the program easier to read and to have the same structure as the dotnet code
