+++ JavaClient/src/main/java/CallHeimdallApi.java
@@ -1,37 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonParser;
-import com.microsoft.aad.msal4j.ClientCredentialFactory;
-import com.microsoft.aad.msal4j.ClientCredentialParameters;
-import com.microsoft.aad.msal4j.ConfidentialClientApplication;
-import com.microsoft.aad.msal4j.IAuthenticationResult;
-import com.nimbusds.oauth2.sdk.http.HTTPResponse;
-import java.io.BufferedReader;
-import java.io.ByteArrayInputStream;
+import Clients.AccessTokenClient;
+
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
-import java.net.URLEncoder;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.security.KeyFactory;
-import java.security.PrivateKey;
-import java.security.cert.CertificateFactory;
-import java.security.cert.X509Certificate;
-import java.security.spec.PKCS8EncodedKeySpec;
-import java.sql.Date;
-import java.text.DateFormat;
-import java.util.Collections;
-import java.util.List;
 import java.util.Properties;
-import java.util.concurrent.CompletableFuture;
-import java.util.stream.Collectors;
+import Services.HeimdallApiService;
+
 
 class CallHeimdallApi {
 
@@ -49,33 +24,13 @@ class CallHeimdallApi {
         try {
             System.out.println("Hello Heimdall!");
 
-            IAuthenticationResult result = getAccessTokenByClientCredentialGrant();
-            System.out.println("Access token = " + result.accessToken());
-            String accessToken = result.accessToken();
+            AccessTokenClient client = new AccessTokenClient(keyPath, certPath, clientId, authority, scope);
+            String accessToken = client.getAccessToken();
+            System.out.println("Access token = " + accessToken);
 
+            HeimdallApiService service = new HeimdallApiService(accessToken, apiUrl);
+            service.Run();
 
-            List<String> lineNames = getLineNames(accessToken);
-           
-            if(lineNames.size() <= 0) {
-                System.out.println("Didn't find any lines");
-            }
-            else{
-                String chosenLine = lineNames.get(0);
-                System.out.println("Requesting data for line with name: " + chosenLine);
-
-                List<AggregatedMeasurement> measurements = getAggregatedMeasurementsForLine(accessToken, chosenLine);
-                for(AggregatedMeasurement measurement : measurements) {
-                    System.out.println(measurement.toString());
-                }
-                System.out.println("Measurements found in the last week: " + measurements.size());                
-                
-                List<DynamicLineRating> dynamicLineRatings = getDynamicLineRatingsForLine(accessToken, chosenLine, DLRType.HP);
-                for(DynamicLineRating dynamicLineRating : dynamicLineRatings) {
-                    System.out.println(dynamicLineRating.toString());
-                }
-                System.out.println("Dynamic line ratings found in the last week: " + measurements.size());
-            }
-           
             System.out.println("Press any key to exit ...");
             System.in.read();
 
@@ -84,234 +39,6 @@ class CallHeimdallApi {
             System.out.println("Exception message - " + ex.getMessage());
             throw ex;
         }
-    }
-
-    private static IAuthenticationResult getAccessTokenByClientCredentialGrant() throws Exception {
-
-        PKCS8EncodedKeySpec spec = new PKCS8EncodedKeySpec(Files.readAllBytes(Paths.get(keyPath)));
-        PrivateKey key = KeyFactory.getInstance("RSA").generatePrivate(spec);
-
-        InputStream certStream = new ByteArrayInputStream(Files.readAllBytes(Paths.get(certPath)));
-        X509Certificate cert = (X509Certificate) CertificateFactory.getInstance("X.509").generateCertificate(certStream);
-
-        ConfidentialClientApplication app = ConfidentialClientApplication.builder(
-                clientId,
-                ClientCredentialFactory.createFromCertificate(key, cert))
-                .authority(authority)
-                .build();
-
-        // With client credentials flows the scope is ALWAYS of the shape "resource/.default", as the
-        // application permissions need to be set statically (in the portal), and then granted by a tenant administrator
-
-        ClientCredentialParameters clientCredentialParam = ClientCredentialParameters.builder(
-                Collections.singleton(scope))
-                .build();
-
-        CompletableFuture<IAuthenticationResult> future = app.acquireToken(clientCredentialParam);
-        return future.get();
-    }
-    private static List<String> getLineNames(String accessToken) throws IOException {
-        StringBuilder urlBuilder = new StringBuilder();
-        urlBuilder.append(apiUrl + "api/beta/lines");
-        System.out.println("Sending request to url: " + urlBuilder.toString());
-        URL url = new URL(urlBuilder.toString());
-        
-        HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-        conn.setRequestMethod("GET");
-        conn.setRequestProperty("Authorization", "Bearer " + accessToken);
-
-        int httpResponseCode = conn.getResponseCode();
-
-        if(httpResponseCode == HTTPResponse.SC_OK) {
-
-            StringBuilder response;
-            try(BufferedReader in = new BufferedReader(
-                    new InputStreamReader(conn.getInputStream()))){
-
-                String inputLine;
-                response = new StringBuilder();
-                while (( inputLine = in.readLine()) != null) {
-                    response.append(inputLine);
-                }
-            }
-            Gson gson = new GsonBuilder().setDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'").create();
-            LineResponse lineResponse = gson.fromJson(response.toString(),LineResponse.class);
-            List<LineDto> lineDtos = lineResponse.data;
-            for(LineDto lineDto : lineDtos) {
-                System.out.println("Found line "+ lineDto.toString() + "\n");
-            }
-
-            List<String> lineNames = lineDtos.stream().map(lineDto -> lineDto.name).collect(Collectors.toList());
-            return lineNames;
-        } else {
-            StringBuilder errorResponse;
-            try(BufferedReader in = new BufferedReader(
-                    new InputStreamReader(conn.getErrorStream()))){
-
-                String inputLine;
-                errorResponse = new StringBuilder();
-                while (( inputLine = in.readLine()) != null) {
-                    errorResponse.append(inputLine);
-                }
-            }
-            System.out.println("Response: " + conn.getResponseMessage());
-        
-            System.out.println(
-                String.format("Connection returned HTTP code: %s with message: %s, details:\n%s",
-                    httpResponseCode, 
-                    conn.getResponseMessage(),
-                    prettifyJsonString(errorResponse.toString())
-                ));
-
-
-            return Collections.emptyList();
-        }
-    }
-    private static List<AggregatedMeasurement> getAggregatedMeasurementsForLine(String accessToken, String lineName) throws IOException {
-
-        Date toDate = new Date(System.currentTimeMillis());
-        Date fromDate = new Date((long) (System.currentTimeMillis() - (7 * 8.64e+7))); // 8.64e+7 = 1 day in milliseconds
-
-        String endpointUrl = apiUrl + "api/beta/aggregated-measurements";
-        StringBuilder paramsBuilder = new StringBuilder();
-        paramsBuilder
-            // You can also use the more detailed format: yyyy-MM-dd'T'HH:mm:ss'Z'
-            .append("?fromDateTime=" + fromDate.toString())
-            .append("&toDateTime=" + toDate.toString())
-            .append("&measurementType=" + MeasurementType.Current)
-            .append("&aggregationType=" + AggregationType.Max)
-            .append("&intervalDuration=" + IntervalDuration.EveryDay)
-            .append("&lineName=" + URLEncoder.encode(lineName, "UTF-8"));
-
-
-        String encodedUrl  = endpointUrl + paramsBuilder.toString();
-        System.out.println("Sending measurement request to url: " + encodedUrl);
-    
-        URL url = new URL(encodedUrl);
-
-        HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-        conn.setRequestMethod("GET");
-        conn.setRequestProperty("Authorization", "Bearer " + accessToken);
-
-        int httpResponseCode = conn.getResponseCode();
-
-        if(httpResponseCode == HTTPResponse.SC_OK) {
-
-            StringBuilder responseBuilder;
-            try(BufferedReader in = new BufferedReader(
-                    new InputStreamReader(conn.getInputStream()))){
-
-                String inputLine;
-                responseBuilder = new StringBuilder();
-                while (( inputLine = in.readLine()) != null) {
-                    responseBuilder.append(inputLine);
-                }
-            }
-            Gson gson = new GsonBuilder().setDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'").create();
-            AggregatedMeasurementsResponse response = gson.fromJson(responseBuilder.toString(),AggregatedMeasurementsResponse.class);
-
-            return response.data;
-        } else {
-            StringBuilder errorResponse;
-            try(BufferedReader in = new BufferedReader(
-                    new InputStreamReader(conn.getErrorStream()))){
-
-                String inputLine;
-                errorResponse = new StringBuilder();
-                while (( inputLine = in.readLine()) != null) {
-                    errorResponse.append(inputLine);
-                }
-            }
-            System.out.println(
-                "Measurement request failed: " + conn.getResponseMessage() + errorResponse);
-        
-            System.out.println(
-                String.format("Connection returned HTTP code: %s with message: %s, details:\n%s",
-                    httpResponseCode, 
-                    conn.getResponseMessage(),
-                    prettifyJsonString(errorResponse.toString())
-                ));
-
-
-            return Collections.emptyList();
-        }
-    }      
-    
-    private static List<DynamicLineRating> getDynamicLineRatingsForLine(String accessToken, String lineName, DLRType dlrType) throws IOException {
-
-        Date toDate = new Date(System.currentTimeMillis());
-        Date fromDate = new Date((long) (System.currentTimeMillis() - (7 * 8.64e+7))); // 8.64e+7 = 1 day in milliseconds
-
-        String endpointUrl = apiUrl + "api/beta/dlr/aggregated-dlr";
-        StringBuilder paramsBuilder = new StringBuilder();
-        paramsBuilder
-            // You can also use the more detailed format: yyyy-MM-dd'T'HH:mm:ss'Z'
-            .append("?fromDateTime=" + fromDate.toString())
-            .append("&toDateTime=" + toDate.toString())
-            .append("&dlrType=" + dlrType)
-            .append("&intervalDuration=" + IntervalDuration.EveryDay)
-            .append("&lineName=" + URLEncoder.encode(lineName, "UTF-8"));
-
-
-        String encodedUrl  = endpointUrl + paramsBuilder.toString();
-        System.out.println("Sending DLR request to url: " + encodedUrl);
-    
-        URL url = new URL(encodedUrl);
-
-        HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-        conn.setRequestMethod("GET");
-        conn.setRequestProperty("Authorization", "Bearer " + accessToken);
-
-        int httpResponseCode = conn.getResponseCode();
-
-        if(httpResponseCode == HTTPResponse.SC_OK) {
-
-            StringBuilder responseBuilder;
-            try(BufferedReader in = new BufferedReader(
-                    new InputStreamReader(conn.getInputStream()))){
-
-                String inputLine;
-                responseBuilder = new StringBuilder();
-                while (( inputLine = in.readLine()) != null) {
-                    responseBuilder.append(inputLine);
-                }
-            }
-            Gson gson = new GsonBuilder().setDateFormat("yyyy-MM-dd'T'HH:mm:ss'Z'").create();
-            DynamicLineRatingResponse response = gson.fromJson(responseBuilder.toString(),DynamicLineRatingResponse.class);
-
-            return response.data;
-        } else {
-            StringBuilder errorResponse;
-            try(BufferedReader in = new BufferedReader(
-                    new InputStreamReader(conn.getErrorStream()))){
-
-                String inputLine;
-                errorResponse = new StringBuilder();
-                while (( inputLine = in.readLine()) != null) {
-                    errorResponse.append(inputLine);
-                }
-            }
-            System.out.println(
-                "DLR request failed: " + conn.getResponseMessage() + errorResponse);
-        
-            System.out.println(
-                String.format("Connection returned HTTP code: %s with message: %s, details:\n%s",
-                    httpResponseCode, 
-                    conn.getResponseMessage(),
-                    prettifyJsonString(errorResponse.toString())
-                ));
-
-
-            return Collections.emptyList();
-        }
-    }    
-
-    private static String prettifyJsonString(String json){
-        JsonParser parser = new JsonParser();
-        Gson gson = new GsonBuilder().setPrettyPrinting().create();
-
-        JsonElement el = parser.parse(json);
-        return gson.toJson(el);
     }
 
     /**
@@ -328,102 +55,8 @@ class CallHeimdallApi {
         certPath = properties.getProperty("CRT_CERTIFICATE_PATH");
         scope = properties.getProperty("SCOPE");
         apiUrl = properties.getProperty("API_URL");
-    }
-
-    public static  class LineResponse {
-        List<LineDto> data;
-        int code;
-        String message;
-        public String toString() {
-            return "LineResponse (" + 
-                "data= " + data +
-                ", code= " + code +
-                ", message= " + message +
-            ")";
-        }
-    }
-    
-    public static  class LineDto {
-        String name;
-        String owner;
-        List<LineSpanDto> lineSpans;
-        public String toString() {
-            return "\nName: " + name +
-                "\nOwner: " + owner +
-                "\nLinespans in line: " + lineSpans.size() + 
-                "\n" + lineSpans;
-        }
-    }
-    public static  class LineSpanDto {
-        String name;
-        public String toString() {
-            return name;
-        }
-    }
-    public static  class AggregatedMeasurementsResponse {
-        List<AggregatedMeasurement> data;
-        String message;
-        public String toString() {
-            return "AggregatedMeasurementResponse (" + 
-                "message= " + message +
-                ", Measurement count= " + data.size();
-        }
-     }
-
-    public static class AggregatedMeasurement {
-        double value;
-        Date intervalStartTime;
-
-        @Override
-        public String toString() {
-            return "Current at " + DateFormat.getInstance().format(intervalStartTime) + ": " + value +" A";
-        }
-    }
-
-    public static  class DynamicLineRatingResponse {
-        List<DynamicLineRating> data;
-        String message;
-        public String toString() {
-            return "DynamicLineRatingResponse (" + 
-                "message= " + message +
-                ", Measurement count= " + data.size();
-        }
-    }
-
-    public static class DynamicLineRating {
-        double ampacity;
-        Date intervalStartTime;
-
-        @Override
-        public String toString() {
-            return "Dynamic line rating at " + DateFormat.getInstance().format(intervalStartTime) + ": " + ampacity +" A";
-        }
-    }
-
-    enum AggregationType {
-        Max,
-        Min,
-        Average
-    }
-    
-    public static class IntervalDuration
-    {
-        public static String Every5Minutes = "PT5M";
-        public static String EveryDay= "P1D";
-
-    }
-    enum MeasurementType
-    {
-        Current,
-        WireTemperature
-    }
-
-    enum DLRType
-    {
-        IEEE,
-        Cigre,
-        HP
-    }
+    }   
 }
+
