+++ JavaClient/src/main/java/Clients/AccessTokenClient.java
@@ -1,0 +1,54 @@
+package Clients;
+
+import java.io.InputStream;
+import java.io.ByteArrayInputStream;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.security.KeyFactory;
+import java.security.PrivateKey;
+import java.security.cert.CertificateFactory;
+import java.security.cert.X509Certificate;
+import java.security.spec.PKCS8EncodedKeySpec;
+import com.microsoft.aad.msal4j.ClientCredentialFactory;
+import com.microsoft.aad.msal4j.ClientCredentialParameters;
+import com.microsoft.aad.msal4j.ConfidentialClientApplication;
+import com.microsoft.aad.msal4j.IAuthenticationResult;
+
+public class AccessTokenClient {
+    private String keyPath;
+    private String certPath;
+    private String clientId;
+    private String authority;
+    private String scope;
+
+    public AccessTokenClient(String keyPath, String certPath, String clientId, String authority, String scope){
+        this.keyPath = keyPath;
+        this.certPath = certPath;
+        this.clientId = clientId;
+        this.authority = authority;
+    }
+
+    public String getAccessToken() throws Exception {
+        PKCS8EncodedKeySpec spec = new PKCS8EncodedKeySpec(Files.readAllBytes(Paths.get(keyPath)));
+        PrivateKey key = KeyFactory.getInstance("RSA").generatePrivate(spec);
+
+        InputStream certStream = new ByteArrayInputStream(Files.readAllBytes(Paths.get(certPath)));
+        X509Certificate cert = (X509Certificate) CertificateFactory.getInstance("X.509").generateCertificate(certStream);
+
+        ConfidentialClientApplication app = ConfidentialClientApplication.builder(
+                clientId,
+                ClientCredentialFactory.createFromCertificate(key, cert))
+                .authority(authority)
+                .build();
+
+        // With client credentials flows the scope is ALWAYS of the shape "resource/.default", as the
+        // application permissions need to be set statically (in the portal), and then granted by a tenant administrator
+
+        ClientCredentialParameters clientCredentialParam = ClientCredentialParameters.builder(
+                Collections.singleton(scope))
+                .build();
+
+        CompletableFuture<IAuthenticationResult> future = app.acquireToken(clientCredentialParam);
+        return future.get().accessToken();
+    }
+}
