+++ JavaClient/src/main/java/Entities/IntervalDuration.java
@@ -1,0 +1,7 @@
+package Entities;
+
+public class IntervalDuration
+{
+    public static String Every5Minutes = "PT5M";
+    public static String EveryDay= "P1D";
+}
