+++ JavaClient/src/main/java/Entities/LineResponse.java
@@ -1,0 +1,17 @@
+package Entities;
+
+import java.util.List;
+
+public class LineResponse {
+    public List<LineDto> data;
+    public int code;
+    public String message;
+    
+    public String toString() {
+        return "LineResponse (" + 
+            "data= " + data +
+            ", code= " + code +
+            ", message= " + message +
+        ")";
+    }
+}
