+++ JavaClient/src/main/java/Entities/LineSpanDto.java
@@ -1,0 +1,9 @@
+package Entities;
+
+public class LineSpanDto {
+    String name;
+    
+    public String toString() {
+        return name;
+    }
+}
