+++ JavaClient/src/main/java/Enums/AggregationType.java
@@ -1,0 +1,7 @@
+package Enums;
+
+public enum AggregationType {
+    Max,
+    Min,
+    Average
+}
