+++ JavaClient/src/main/java/Enums/DLRType.java
@@ -1,0 +1,7 @@
+package Enums;
+
+public enum DLRType {
+    IEEE,
+    Cigre,
+    HP
+}
