+++ JavaClient/src/main/java/Enums/MeasurementType.java
@@ -1,0 +1,6 @@
+package Enums;
+
+public enum MeasurementType {
+    Current,
+    WireTemperature
+}
